[Refactor] HBTIVC의 상단 및 하단 뷰를 별도의 클래스로 분리
