+++ HMOA_iOS/HMOA_iOS/Source/Presentation/Tab/Home/HBTI/HBTI/View/HBTIHomeBottomView.swift
@@ -1,5 +1,5 @@
 //
-//  HBTIViewController.swift
+//  HBTIHomeBottomView.swift
 //  HMOA_iOS
 //
 //  Created by 곽다은 on 7/11/24.
@@ -7,59 +7,9 @@
 
 import UIKit
 
-import SnapKit
-import ReactorKit
-import RxCocoa
-import RxSwift
-import SnapKit
-import Then
+class HBTIHomeBottomView: UIView {
 
-class HBTIViewController: UIViewController, View {
-    
     // MARK: - UI Components
-    
-    private let scrollView = UIScrollView().then{
-        $0.backgroundColor = .black
-    }
-    
-    // 당신의 향BTI는 무엇일까요?
-    private let yourHBTIView = UIView()
-    
-    private let titleLabel = UILabel().then {
-        $0.setLabelUI("당신의 향BTI는 무엇일까요?", font: .pretendard, size: 20, color: .white)
-    }
-    
-    private let descriptionLabel = UILabel().then {
-        $0.setLabelUI("", font: .pretendard, size: 14, color: .white)
-        $0.setTextWithLineHeight(text: "검사를 통해 좋아하는 향료와\n향수까지 알아보세요!", lineHeight: 17)
-        $0.numberOfLines = 2
-    }
-    
-    private let buttonStackView = UIStackView().then {
-        $0.axis = .horizontal
-        $0.spacing = 12
-        $0.alignment = .fill
-        $0.distribution = .fillEqually
-    }
-    
-    private let goToSurveyButton = UIButton().then {
-        $0.setTitle("향BTI\n검사하러 가기", for: .normal)
-        $0.setTitleColor(.black, for: .normal)
-        $0.layer.masksToBounds = true
-        $0.layer.cornerRadius = 5
-        $0.setBackgroundImage(UIImage(named: "goToSurvey"), for: .normal)
-    }
-    
-    private let selectSpiceButton = UIButton().then {
-        $0.setTitle("향료 입력하기\n(주문 후)", for: .normal)
-        $0.setTitleColor(.black, for: .normal)
-        $0.layer.masksToBounds = true
-        $0.layer.cornerRadius = 5
-        $0.setBackgroundImage(UIImage(named: "selectSpice"), for: .normal)
-    }
-    
-    // 향BTI란?
-    private let introView = UIView()
     
     private let introTitleView = UIView()
     
@@ -110,58 +60,26 @@ class HBTIViewController: UIViewController, View {
         """
     )
     
-    // MARK: - Properties
+    // MARK: - Init
     
-    var disposeBag = DisposeBag()
-    
-    // MARK: - LifeCycle
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
+    override init(frame: CGRect) {
+        super.init(frame: frame)
         
-        setUI()
         setAddView()
         setConstraints()
     }
     
-    // MARK: - Bind
-    
-    func bind(reactor: HBTIReactor) {
-        
-        // MARK: Action
-        
-        
-        // MARK: State
-        
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
     
-    // MARK: - Functions
+    // MARK: - Set UI
     
-    private func setUI() {
-        setClearBackNaviBar("향BTI", .white)
-        
-    }
-    
-    // MARK: Add Views
     private func setAddView() {
-        view.addSubview(scrollView)
-        
-        [yourHBTIView, 
-         introView
-        ].forEach { scrollView.addSubview($0) }
-        
-        [titleLabel,
-         descriptionLabel,
-         buttonStackView
-        ].forEach { yourHBTIView.addSubview($0) }
-        
-        [goToSurveyButton, selectSpiceButton
-        ].forEach { buttonStackView.addArrangedSubview($0)}
-        
         [introTitleView,
          introDesctiptionLabel,
          introStackView
-        ].forEach{ introView.addSubview($0) }
+        ].forEach{ self.addSubview($0) }
         
         [logoImageView, introTitleLabel
         ].forEach { introTitleView.addSubview($0) }
@@ -172,38 +90,7 @@ class HBTIViewController: UIViewController, View {
         ].forEach { introStackView.addArrangedSubview($0) }
     }
     
-    // MARK: Set Constraints
     private func setConstraints() {
-        scrollView.snp.makeConstraints { make in
-            make.edges.equalToSuperview()
-        }
-        
-        yourHBTIView.snp.makeConstraints { make in
-            make.top.equalToSuperview().inset(20)
-            make.leading.trailing.equalToSuperview().inset(16)
-        }
-        
-        titleLabel.snp.makeConstraints { make in
-            make.top.leading.trailing.equalToSuperview()
-        }
-        
-        descriptionLabel.snp.makeConstraints { make in
-            make.top.equalTo(titleLabel.snp.bottom).offset(12)
-            make.leading.trailing.equalToSuperview()
-        }
-        
-        buttonStackView.snp.makeConstraints { make in
-            make.top.equalTo(descriptionLabel.snp.bottom).offset(20)
-            make.leading.trailing.bottom.equalToSuperview()
-            make.height.equalTo(107)
-        }
-        
-        introView.snp.makeConstraints { make in
-            make.top.equalTo(yourHBTIView.snp.bottom).offset(32)
-            make.leading.trailing.equalToSuperview().inset(16)
-            make.bottom.equalToSuperview().inset(20)
-        }
-        
         introTitleView.snp.makeConstraints { make in
             make.top.leading.trailing.equalToSuperview()
         }
@@ -230,6 +117,7 @@ class HBTIViewController: UIViewController, View {
             make.width.lessThanOrEqualTo(UIScreen.main.bounds.width - 32)
         }
     }
+
 }
 
 // MARK: UIIntroView
